My two cents to improve the animation
Changing the ValueAnimator starting point to 0.1f gives the impression
that the waves are starting from the bottom of the drawable, rather
than starting from underneath the drawable. It makes it feel more real.
I’ve also changed the repeat mode to REVERSE as this gives a much nicer
animation for indeterminate animations.